updated type for id columns from integer to bigint

+++ Backend-Bot-master/app/models/car.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, TIMESTAMP, func, Text
+from sqlalchemy import BigInteger, String, TIMESTAMP, func, Text
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db import Base
 from sqlalchemy import ForeignKey
@@ -7,8 +7,8 @@ from sqlalchemy.orm import relationship
 class Car(Base):
     __tablename__ = 'cars'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    driver_profile_id: Mapped[int] = mapped_column(Integer, ForeignKey('driver_profiles.id'), nullable=False)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    driver_profile_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('driver_profiles.id'), nullable=False)
     model: Mapped[str] = mapped_column(String(100), nullable=True)
     number: Mapped[str] = mapped_column(String(100), nullable=True)
     region: Mapped[str] = mapped_column(String(20), nullable=True)
